requirements.txt
fastapi
yfinance
requests
python-dotenv
langchain
langchain-core
langchain-google-genai
langchain-openai
google-genai
pydantic
uvicorn[standard]
pandas
langchain_community
redis
pyyaml